Using labels instead of placeholders
I have set the element_placeholder argument to inform the label argument instead of the placeholder argument in textAreaInput to conform to accessibility standards. Set a generic "Choose from list or insert new values" or "Choose from list" placeholder for the drop down menus. Removed the boldness of the input labels to distinguish them from the other headings on the pages.